fix(shared): rename try[On]BeforeUnMount
